Chuangmi Plug M1 merged.
