+++ miio/__init__.py
@@ -3,7 +3,6 @@ from miio.protocol import Message, Utils
 from miio.vacuumcontainers import (VacuumStatus, ConsumableStatus, DNDStatus,
                                    CleaningDetails, CleaningSummary, Timer, )
 from miio.vacuum import Vacuum, VacuumException
-from miio.plug import Plug
 from miio.chuangmi_plug import ChuangmiPlug
 from miio.airpurifier import AirPurifier
 from miio.airhumidifier import AirHumidifier
