+++ miio/tests/test_chuangmi_plug.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 from miio import ChuangmiPlug
 from miio.chuangmi_plug import (ChuangmiPlugStatus, MODEL_CHUANGMI_PLUG_V1,
-                                MODEL_CHUANGMI_PLUG_V3, )
+                                MODEL_CHUANGMI_PLUG_V3, MODEL_CHUANGMI_PLUG_M1)
 from .dummies import DummyDevice
 import pytest
 
@@ -177,3 +177,54 @@ class TestChuangmiPlugV3(TestCase):
 
         self.device.set_wifi_led(False)
         assert wifi_led() is False
+
+
+class DummyChuangmiPlugM1(DummyDevice, ChuangmiPlug):
+    def __init__(self, *args, **kwargs):
+        self.model = MODEL_CHUANGMI_PLUG_M1
+        self.state = {
+            'power': 'on',
+            'temperature': 32,
+        }
+        self.return_values = {
+            'get_prop': self._get_state,
+            'set_power': lambda x: self._set_state("power", x),
+        }
+        super().__init__(args, kwargs)
+
+
+@pytest.fixture(scope="class")
+def chuangmiplugm1(request):
+    request.cls.device = DummyChuangmiPlugM1()
+    # TODO add ability to test on a real device
+
+
+@pytest.mark.usefixtures("chuangmiplugm1")
+class TestChuangmiPlugM1(TestCase):
+    def is_on(self):
+        return self.device.status().is_on
+
+    def state(self):
+        return self.device.status()
+
+    def test_on(self):
+        self.device.off()  # ensure off
+        assert self.is_on() is False
+
+        self.device.on()
+        assert self.is_on() is True
+
+    def test_off(self):
+        self.device.on()  # ensure on
+        assert self.is_on() is True
+
+        self.device.off()
+        assert self.is_on() is False
+
+    def test_status(self):
+        self.device._reset_state()
+
+        assert repr(self.state()) == repr(ChuangmiPlugStatus(self.device.start_state))
+
+        assert self.is_on() is True
+        assert self.state().temperature == self.device.start_state["temperature"]
